server/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.1.2
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
joblib==1.1.0
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.3
Pillow==9.1.0
python-dotenv==0.20.0
pytz==2022.1
regex==2022.3.15
scikit-learn==1.0.2
scipy==1.8.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.1
